- Implemented FeedUsecase.kt

+++ kotlin_app/src/main/java/life/league/challenge/kotlin/data/network/model/PostApiModel.kt
@@ -3,8 +3,8 @@ package life.league.challenge.kotlin.data.network.model
 import com.google.gson.annotations.SerializedName
 
 data class PostApiModel(
-        @SerializedName("id") val id: Int,
+        @SerializedName("id") val id: Long,
         @SerializedName("title") val title: String,
         @SerializedName("body") val body: String,
-        @SerializedName("userId") val userId: Int
+        @SerializedName("userId") val userId: Long
 )
